feat(matrix): Matrix now generates custom pin labels
